Add RawDataPacketBuilder

+++ include/lldt/validated_source_frame_view.hpp
@@ -9,7 +9,7 @@ namespace transport
     struct ValidatedSourceFrameView
     {
         const std::byte* data{nullptr};
-        std::size_t frame_size{};
+        std::uint32_t frame_size{};
         std::uint64_t source_seq_id{};
         bool begins_after_source_gap{false};
     };
